+++ src/deemon/objects/seq/each.c
@@ -440,7 +440,7 @@ se_hasitem(SeqEachBase *self, DeeObject *index) {
 	Dee_ssize_t foreach_status;
 	foreach_status = DeeObject_InvokeMethodHint(seq_operator_foreach, self->se_seq,
 	                                            &se_hasitem_foreach_cb, index);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 PRIVATE WUNUSED NONNULL((2)) Dee_ssize_t DCALL
@@ -470,7 +470,7 @@ se_hasitem_index(SeqEachBase *self, size_t index) {
 	foreach_status = DeeObject_InvokeMethodHint(seq_operator_foreach, self->se_seq,
 	                                            &se_hasitem_index_foreach_cb,
 	                                            (void *)(uintptr_t)index);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 struct se_bounditem_string_hash_foreach_data {
@@ -519,7 +519,7 @@ se_hasitem_string_hash(SeqEachBase *self, char const *key, Dee_hash_t hash) {
 	foreach_status = DeeObject_InvokeMethodHint(seq_operator_foreach, self->se_seq,
 	                                                &se_hasitem_string_hash_foreach_cb,
 	                                                &data);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 struct se_bounditem_string_len_hash_foreach_data {
@@ -573,7 +573,7 @@ se_hasitem_string_len_hash(SeqEachBase *self, char const *key, size_t keylen, De
 	foreach_status = DeeObject_InvokeMethodHint(seq_operator_foreach, self->se_seq,
 	                                            &se_hasitem_string_len_hash_foreach_cb,
 	                                            &data);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 
@@ -735,7 +735,7 @@ se_hasattr(SeqEachBase *self, DeeObject *attr) {
 	Dee_ssize_t foreach_status;
 	foreach_status = DeeObject_InvokeMethodHint(seq_operator_foreach, self->se_seq,
 	                                            &se_hasattr_foreach_cb, attr);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 struct se_boundattr_string_hash_foreach_data {
@@ -783,7 +783,7 @@ se_hasattr_string_hash(SeqEachBase *self, char const *attr, Dee_hash_t hash) {
 	foreach_status = DeeObject_InvokeMethodHint(seq_operator_foreach, self->se_seq,
 	                                            &se_hasattr_string_hash_foreach_cb,
 	                                            &data);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 struct se_boundattr_string_len_hash_foreach_data {
@@ -837,7 +837,7 @@ se_hasattr_string_len_hash(SeqEachBase *self, char const *attr, size_t attrlen, 
 	foreach_status = DeeObject_InvokeMethodHint(seq_operator_foreach, self->se_seq,
 	                                            &se_hasattr_string_len_hash_foreach_cb,
 	                                            &data);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 
@@ -2897,7 +2897,7 @@ PRIVATE WUNUSED NONNULL((1, 2)) int DCALL
 seo_hasattr(SeqEachOperator *self, DeeObject *attr) {
 	Dee_ssize_t foreach_status;
 	foreach_status = seo_foreach(self, &se_hasattr_foreach_cb, attr);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 PRIVATE WUNUSED NONNULL((1, 2)) int DCALL
@@ -2917,7 +2917,7 @@ seo_hasattr_string_hash(SeqEachOperator *self, char const *attr, Dee_hash_t hash
 	data.ssbashfd_attr = attr;
 	data.ssbashfd_hash = hash;
 	foreach_status = seo_foreach(self, &se_hasattr_string_hash_foreach_cb, &data);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 PRIVATE WUNUSED NONNULL((1, 2)) int DCALL
@@ -2939,7 +2939,7 @@ seo_hasattr_string_len_hash(SeqEachOperator *self, char const *attr, size_t attr
 	data.ssbaslhfd_attrlen = attrlen;
 	data.ssbaslhfd_hash    = hash;
 	foreach_status = seo_foreach(self, &se_hasattr_string_len_hash_foreach_cb, &data);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 
@@ -3179,7 +3179,7 @@ PRIVATE WUNUSED NONNULL((1, 2)) int DCALL
 seo_operator_hasitem(SeqEachOperator *self, DeeObject *index) {
 	Dee_ssize_t foreach_status;
 	foreach_status = seo_foreach(self, &se_hasitem_foreach_cb, index);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 #define seo_operator_bounditem_index_PTR &seo_operator_bounditem_index
@@ -3197,7 +3197,7 @@ seo_operator_hasitem_index(SeqEachOperator *self, size_t index) {
 	Dee_ssize_t foreach_status;
 	foreach_status = seo_foreach(self, &se_hasitem_index_foreach_cb,
 	                             (void *)(uintptr_t)index);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 #define seo_operator_bounditem_string_hash_PTR &seo_operator_bounditem_string_hash
@@ -3219,7 +3219,7 @@ seo_operator_hasitem_string_hash(SeqEachOperator *self, char const *key, Dee_has
 	data.ssbishfd_key  = key;
 	data.ssbishfd_hash = hash;
 	foreach_status = seo_foreach(self, &se_hasitem_string_hash_foreach_cb, &data);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 #define seo_operator_bounditem_string_len_hash_PTR &seo_operator_bounditem_string_len_hash
@@ -3243,7 +3243,7 @@ seo_operator_hasitem_string_len_hash(SeqEachOperator *self, char const *key, siz
 	data.ssbislhfd_keylen = keylen;
 	data.ssbislhfd_hash   = hash;
 	foreach_status = seo_foreach(self, &se_hasitem_string_len_hash_foreach_cb, &data);
-	return Dee_HAS_FROM_eM1_nM2_y0(foreach_status);
+	return Dee_HAS_FROM_eM1_nM2_y0((int)foreach_status);
 }
 
 #define seo_operator_sizeob_PTR                     &sew_sizeob
